requirements.txt
boto3==1.24.82
botocore==1.27.82
jsonschema==4.17.3
python-dotenv==0.21.0
